Added the buttons for Profile and the buttons from Home for different…

+++ Controllers/ProfilesController.cs
@@ -1,4 +1,5 @@
 ﻿using DAW_social_platform.Models;
+using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -12,6 +13,7 @@ namespace DAW_social_platform.Controllers
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: Profiles
+        [Authorize(Roles = "User, Admin")]
         public ActionResult Index()
         {
             if (TempData.ContainsKey("message"))
@@ -24,11 +26,19 @@ namespace DAW_social_platform.Controllers
             return View();
         }
 
+        [Authorize(Roles = "User,Admin")]
         public ActionResult New()
         {
-            return View();
+            var userId = User.Identity.GetUserId();
+            Profile profile = (from pr in db.Profiles
+                               where pr.UserId == userId
+                               select pr).FirstOrDefault();
+            if (profile == null)
+                return View();
+            else return RedirectToAction("Index", "Home");
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public ActionResult New(Profile profile)
         {
@@ -36,12 +46,13 @@ namespace DAW_social_platform.Controllers
             {
                 var date = profile.BirthDate;
                 profile.BirthDate = Convert.ToDateTime(date);
+                profile.UserId = User.Identity.GetUserId();
                 if (ModelState.IsValid)
                 {
                     db.Profiles.Add(profile);
                     db.SaveChanges();
                     TempData["message"] = "Profilul a fost creat!";
-                    return RedirectToAction("Index");
+                    return RedirectToAction("Index", "Manage");
                 }
                 else
                 {
@@ -54,31 +65,57 @@ namespace DAW_social_platform.Controllers
             }
         }
 
+        [Authorize(Roles = "User,Admin")]
         public ActionResult Edit(int id)
         {
             var profile = db.Profiles.Find(id);
-            return View(profile);
+            if (profile != null)
+            {
+                if (profile.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+                    return View(profile);
+                else
+                {
+                    TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui profil care nu va apartine";
+                    return RedirectToAction("Index", "Home");
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPut]
         public ActionResult Edit(int id, Profile requestProfile)
         {
             try
             {
-                Profile profile = db.Profiles.Find(id);
-
-                if (TryUpdateModel(profile))
+                if (ModelState.IsValid)
                 {
-                    profile.Description = requestProfile.Description;
-                    profile.BirthDate = Convert.ToDateTime(requestProfile.BirthDate);
-                    profile.Hobbies = requestProfile.Hobbies;
-                    profile.Status = requestProfile.Status;
-                    profile.UserId = requestProfile.UserId;
-                    db.SaveChanges();
-                    TempData["message"] = "Profilul a fost modificat!";
-                    return RedirectToAction("Index");
+                    Profile profile = db.Profiles.Find(id);
+
+                    if (profile.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+                    {
+                        if (TryUpdateModel(profile))
+                        {
+                            profile.Description = requestProfile.Description;
+                            profile.BirthDate = Convert.ToDateTime(requestProfile.BirthDate);
+                            profile.Hobbies = requestProfile.Hobbies;
+                            profile.Status = requestProfile.Status;
+                            profile.UserId = requestProfile.UserId;
+                            db.SaveChanges();
+                            TempData["message"] = "Profilul a fost modificat!";
+                        }
+                        return RedirectToAction("Index", "Manage");
+                    }
+                    else
+                    {
+                        TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui profil care nu va apartine";
+                        return RedirectToAction("Index", "Home");
+                    }
                 }
-                return View(requestProfile);
+                else
+                {
+                    return View(requestProfile);
+                }
             }
             catch (Exception e)
             {
@@ -86,15 +123,23 @@ namespace DAW_social_platform.Controllers
             }
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpDelete]
         public ActionResult Delete(int id)
         {
             Profile profile = db.Profiles.Find(id);
-            db.Profiles.Remove(profile);
-            TempData["message"] = "Profilul a fost sters!";
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (profile.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                db.Profiles.Remove(profile);
+                TempData["message"] = "Profilul a fost sters!";
+                db.SaveChanges();
+                return RedirectToAction("Index", "Home");
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa stergeti un profil care nu va apartine";
+                return RedirectToAction("Index");
+            }
         }
-
     }
 }
